+++ cmd/wpexportjson/main.go
@@ -27,7 +27,7 @@ import (
 
 // Version information - set during build
 var (
-	Version   = "1.5.0"
+	Version   = "1.6.0"
 	BuildDate = "unknown"
 )